api/services/analytics_service.py
"""
Analytics Service - Advanced financial analytics and insights
"""

from typing import Dict, List, Any, Optional
from datetime import datetime, timedelta
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler


class AnalyticsService:
    def __init__(self, db_client, user_id: str):
        self.db = db_client
        self.user_id = user_id
    
    async def get_complete_dashboard(self, request) -> Dict[str, Any]:
        """Generate complete financial dashboard"""
        return {
            "period_info": {
                "period": request.period,
                "start_date": request.start_date,
                "end_date": request.end_date
            },
            "financial_health": await self.calculate_financial_health(),
            "expense_analytics": await self.get_expense_analytics(request),
            "income_analytics": await self.get_income_analytics(request),
            "budget_performance": await self.get_budget_performance(),
            "savings_analytics": await self.get_savings_analytics(),
            "cash_flow_predictions": await self.predict_cash_flow(3) if request.include_predictions else None,
            "insights": ["Sample insight 1", "Sample insight 2"],
            "alerts": []
        }
    
    async def calculate_financial_health(self) -> Dict[str, Any]:
        """Calculate comprehensive financial health score"""
        # Mock implementation - replace with real calculations
        return {
            "overall_score": 78,
            "savings_ratio": 0.15,
            "debt_to_income_ratio": 0.25,
            "spending_consistency": 0.82,
            "emergency_fund_months": 3.5,
            "recommendations": []
        }
    
    async def get_expense_analytics(self, request) -> Dict[str, Any]:
        """Get detailed expense analytics"""
        return {
            "period": request.period,
            "start_date": request.start_date or datetime.now().date(),
            "end_date": request.end_date or datetime.now().date(),
            "total_expenses": {
                "amount": 2500.0,
                "currency": "USD"
            },
            "categories": [],
            "top_merchants": [],
            "unusual_spending": []
        }
    
    async def get_income_analytics(self, request) -> Dict[str, Any]:
        """Get detailed income analytics"""
        return {
            "period": request.period,
            "start_date": request.start_date or datetime.now().date(),
            "end_date": request.end_date or datetime.now().date(),
            "total_income": {
                "amount": 3500.0,
                "currency": "USD"
            },
            "sources": [],
            "growth_rate": 0.05,
            "stability_score": 0.85
        }
    
    async def get_budget_performance(self) -> List[Dict[str, Any]]:
        """Get budget performance data"""
        return []
    
    async def get_savings_analytics(self) -> Dict[str, Any]:
        """Get savings analytics"""
        return {
            "total_saved": {
                "amount": 5000.0,
                "currency": "USD"
            },
            "savings_rate": 0.15,
            "monthly_average": {
                "amount": 400.0,
                "currency": "USD"
            },
            "savings_goals_progress": [],
            "optimal_savings_amount": {
                "amount": 500.0,
                "currency": "USD"
            }
        }
    
    async def analyze_spending_patterns(self, period: str, categories: Optional[List[str]], include_predictions: bool) -> Dict[str, Any]:
        """Analyze spending patterns with ML"""
        # Mock implementation
        return {
            "patterns": [],
            "anomalies": [],
            "predictions": [] if include_predictions else None
        }
    
    async def predict_cash_flow(self, months_ahead: int, include_scenarios: bool = False) -> List[Dict[str, Any]]:
        """Predict future cash flow"""
        predictions = []
        for i in range(months_ahead):
            future_date = datetime.now().date() + timedelta(days=30 * (i + 1))
            predictions.append({
                "date": future_date.isoformat(),
                "predicted_income": {
                    "amount": 3500.0,
                    "currency": "USD"
                },
                "predicted_expenses": {
                    "amount": 2800.0,
                    "currency": "USD"
                },
                "predicted_balance": {
                    "amount": 700.0,
                    "currency": "USD"
                },
                "confidence_score": 0.85
            })
        return predictions
    
    async def analyze_trend(self, metric: str, period: str, lookback_months: int) -> Dict[str, Any]:
        """Analyze trends for financial metrics"""
        return {
            "metric_name": metric,
            "period": period,
            "data_points": [],
            "trend_direction": "stable",
            "trend_strength": 0.6,
            "seasonal_patterns": []
        }
    
    async def compare_periods(self, comparison_type: str, current_period_start: Optional[datetime], current_period_end: Optional[datetime]) -> Dict[str, Any]:
        """Compare different periods"""
        return {
            "comparison_type": comparison_type,
            "current_period": {},
            "previous_period": {},
            "change_percentage": 5.2,
            "change_amount": {
                "amount": 150.0,
                "currency": "USD"
            },
            "significant_changes": []
        }
    
    async def get_user_financial_summary(self) -> Dict[str, Any]:
        """Get user's financial data summary for AI processing"""
        return {
            "total_balance": 10000.0,
            "monthly_income": 3500.0,
            "monthly_expenses": 2800.0,
            "categories": [],
            "recent_transactions": []
        }
    
    async def export_analytics(self, format: str, include_insights: bool, include_predictions: bool) -> Dict[str, Any]:
        """Export analytics in various formats"""
        return {
            "download_url": f"/api/v1/exports/analytics_{self.user_id}.{format}",
            "expires_at": (datetime.utcnow() + timedelta(hours=24)).isoformat(),
            "file_size": "2.5MB"
        }