api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.0.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
reportlab==4.0.7
jinja2==3.1.2
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
asyncpg==0.29.0
alembic==1.13.0
sqlalchemy==2.0.23
pydantic-settings==2.1.0